+++ bin/dojo.js
@@ -1,4 +1,4 @@
 #! /usr/bin/env node
 'use strict'
 
-require('../_build/lib/index');
+require('../_build/index');
