+++ src/logger.ts
@@ -1,4 +1,4 @@
-import { add, remove, transports } from 'winston';
+import { add, remove, transports, log } from 'winston';
 
 export interface VerboseOptions {
 	verbose?: boolean;
@@ -10,4 +10,12 @@ export default function setupLogger(verbose: boolean = false) {
 		showLevel: false,
 		level: verbose ? 'verbose' : 'info'
 	});
+}
+
+export function verbose(message: string) {
+	log('verbose', message);
+}
+
+export function info(message: string) {
+	log('info', message);
 }
