+++ tsconfig.json
@@ -11,7 +11,6 @@
 	},
 	"include": [
 		"./typings/index.d.ts",
-		"./src/util/*.ts",
-		"./src/lib/*.ts"
+		"./src/*.ts"
 	]
 }
